１話当たりの最大フレーム数を2^32まで対応できるように変更。動画のインデックス作成方法を追記。

+++ src/ImageSearchEngine/ImageInfo.cs
@@ -8,6 +8,6 @@ namespace Mpga.ImageSearchEngine
         public ulong Hash;
         public UInt16 TitleId;
         public UInt16 EpisodeId;
-        public UInt16 Frame;
+        public UInt32 Frame;
     }
 }
